+++ Mirror/Weaver/Mirror.Weaver.csproj
@@ -69,6 +69,7 @@
     <Compile Include="MessageClassProcessor.cs" />
     <Compile Include="MonoBehaviourProcessor.cs" />
     <Compile Include="NetworkBehaviourCommandProcessor.cs" />
+    <Compile Include="NetworkBehaviourRpcProcessor.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="SyncListStructProcessor.cs" />
     <Compile Include="NetworkBehaviourProcessor.cs" />
