+++ Mirror/Weaver/NetworkBehaviourProcessor.cs
@@ -28,7 +28,6 @@ namespace Mirror.Weaver
         const int k_SyncVarLimit = 64; // ulong = 64 bytes
         readonly TypeDefinition m_td;
 
-        const string k_RpcPrefix = "InvokeRpc";
         const string k_TargetRpcPrefix = "InvokeTargetRpc";
 
         public NetworkBehaviourProcessor(TypeDefinition td)
@@ -824,7 +823,7 @@ namespace Mirror.Weaver
 
         MethodDefinition ProcessTargetRpcInvoke(MethodDefinition md)
         {
-            MethodDefinition rpc = new MethodDefinition(k_RpcPrefix + md.Name, MethodAttributes.Family |
+            MethodDefinition rpc = new MethodDefinition(NetworkBehaviourRpcProcessor.k_RpcPrefix + md.Name, MethodAttributes.Family |
                     MethodAttributes.Static |
                     MethodAttributes.HideBySig,
                     Weaver.voidType);
@@ -842,34 +841,6 @@ namespace Mirror.Weaver
             rpcWorker.Append(rpcWorker.Create(OpCodes.Call, Weaver.ReadyConnectionReference));
 
             if (!ProcessNetworkReaderParameters(m_td, md, rpcWorker, true))
-                return null;
-
-            // invoke actual command function
-            rpcWorker.Append(rpcWorker.Create(OpCodes.Callvirt, md));
-            rpcWorker.Append(rpcWorker.Create(OpCodes.Ret));
-
-            AddInvokeParameters(rpc.Parameters);
-
-            return rpc;
-        }
-
-        MethodDefinition ProcessRpcInvoke(MethodDefinition md)
-        {
-            MethodDefinition rpc = new MethodDefinition(k_RpcPrefix + md.Name, MethodAttributes.Family |
-                    MethodAttributes.Static |
-                    MethodAttributes.HideBySig,
-                    Weaver.voidType);
-
-            ILProcessor rpcWorker = rpc.Body.GetILProcessor();
-            Instruction label = rpcWorker.Create(OpCodes.Nop);
-
-            WriteClientActiveCheck(rpcWorker, md.Name, label, "RPC");
-
-            // setup for reader
-            rpcWorker.Append(rpcWorker.Create(OpCodes.Ldarg_0));
-            rpcWorker.Append(rpcWorker.Create(OpCodes.Castclass, m_td));
-
-            if (!ProcessNetworkReaderParameters(m_td, md, rpcWorker, false))
                 return null;
 
             // invoke actual command function
@@ -946,62 +917,6 @@ namespace Mirror.Weaver
             rpcWorker.Append(rpcWorker.Create(OpCodes.Ldloc_0)); // writer
             rpcWorker.Append(rpcWorker.Create(OpCodes.Ldc_I4, GetChannelId(ca)));
             rpcWorker.Append(rpcWorker.Create(OpCodes.Callvirt, Weaver.sendTargetRpcInternal));
-
-            rpcWorker.Append(rpcWorker.Create(OpCodes.Ret));
-
-            return rpc;
-        }
-
-        /* generates code like:
-        public void CallRpcTest (int param)
-        {
-            if (!NetworkServer.get_active ()) {
-                Debug.LogError ((object)"RPC Function RpcTest called on client.");
-            } else {
-                NetworkWriter writer = new NetworkWriter ();
-                writer.WritePackedUInt32((uint)param);
-                base.SendRPCInternal("RpcTest", writer, 0);
-            }
-        }
-        */
-        MethodDefinition ProcessRpcCall(MethodDefinition md, CustomAttribute ca)
-        {
-            MethodDefinition rpc = new MethodDefinition("Call" +  md.Name, MethodAttributes.Public |
-                    MethodAttributes.HideBySig,
-                    Weaver.voidType);
-
-            // add paramters
-            foreach (ParameterDefinition pd in md.Parameters)
-            {
-                rpc.Parameters.Add(new ParameterDefinition(pd.Name, ParameterAttributes.None, pd.ParameterType));
-            }
-
-            ILProcessor rpcWorker = rpc.Body.GetILProcessor();
-            Instruction label = rpcWorker.Create(OpCodes.Nop);
-
-            WriteSetupLocals(rpcWorker);
-
-            WriteServerActiveCheck(rpcWorker, md.Name, label, "RPC Function");
-
-            WriteCreateWriter(rpcWorker);
-
-            // write all the arguments that the user passed to the Rpc call
-            if (!WriteArguments(rpcWorker, md, "RPC", false))
-                return null;
-
-            var rpcName = md.Name;
-            int index = rpcName.IndexOf(k_RpcPrefix);
-            if (index > -1)
-            {
-                rpcName = rpcName.Substring(k_RpcPrefix.Length);
-            }
-
-            // invoke SendInternal and return
-            rpcWorker.Append(rpcWorker.Create(OpCodes.Ldarg_0)); // this
-            rpcWorker.Append(rpcWorker.Create(OpCodes.Ldstr, rpcName));
-            rpcWorker.Append(rpcWorker.Create(OpCodes.Ldloc_0)); // writer
-            rpcWorker.Append(rpcWorker.Create(OpCodes.Ldc_I4, GetChannelId(ca)));
-            rpcWorker.Append(rpcWorker.Create(OpCodes.Callvirt, Weaver.sendRpcInternal));
 
             rpcWorker.Append(rpcWorker.Create(OpCodes.Ret));
 
@@ -1270,13 +1185,13 @@ namespace Mirror.Weaver
                         names.Add(md.Name);
                         m_Rpcs.Add(md);
 
-                        MethodDefinition rpcFunc = ProcessRpcInvoke(md);
+                        MethodDefinition rpcFunc = NetworkBehaviourRpcProcessor.ProcessRpcInvoke(m_td, md);
                         if (rpcFunc != null)
                         {
                             m_RpcInvocationFuncs.Add(rpcFunc);
                         }
 
-                        MethodDefinition rpcCallFunc = ProcessRpcCall(md, ca);
+                        MethodDefinition rpcCallFunc = NetworkBehaviourRpcProcessor.ProcessRpcCall(md, ca);
                         if (rpcCallFunc != null)
                         {
                             m_RpcCallFuncs.Add(rpcCallFunc);
